% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteredSignal.R
\docType{class}
\name{ClusteredSignal-class}
\alias{ClusteredSignal-class}
\title{ClusteredSignal}
\description{
ClusteredSignal
}
\section{Slots}{

\describe{
\item{\code{signal_data}}{data.table.}

\item{\code{query_gr}}{GRanges.}

\item{\code{signal_var}}{character.}

\item{\code{signal_var.within}}{character.}

\item{\code{facet_var}}{character.}

\item{\code{extra_var}}{character.}

\item{\code{manual_assigned}}{list.}
}}

