% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteredSignal.R
\name{ClusteredSignal.null}
\alias{ClusteredSignal.null}
\title{ClusteredSignal.null}
\usage{
ClusteredSignal.null()
}
\value{
An empty ClusteredSignal object
}
\description{
ClusteredSignal.null
}
\examples{
ClusteredSignal.null()
}
