% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteredSignal_TSNE.R
\docType{class}
\name{ClusteredSignal_TSNE-class}
\alias{ClusteredSignal_TSNE-class}
\title{ClusteredSignal_TSNE}
\description{
ClusteredSignal_TSNE
}
\section{Slots}{

\describe{
\item{\code{signal_data}}{}

\item{\code{query_gr}}{}

\item{\code{signal_var}}{}

\item{\code{facet_var}}{}

\item{\code{extra_var}}{}
}}

