% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteredSignal_TSNE.R
\name{ClusteredSignal_TSNE.fromConfig}
\alias{ClusteredSignal_TSNE.fromConfig}
\title{ClusteredSignal_TSNE.fromConfig}
\usage{
ClusteredSignal_TSNE.fromConfig(
  signal_config,
  query_gr,
  manual_assigned = list(),
  nclust = 6,
  facet_var = "name_split",
  extra_var = character(),
  bfc = new_cache()
)
}
\description{
ClusteredSignal_TSNE.fromConfig
}
\examples{
options(mc.cores = 10)
set.seed(0)
feature_config_file = system.file(package = "ssvQC", "extdata/ssvQC_peak_config.csv")
feature_config = QcConfigFeatures.parse(feature_config_file, process_features= TRUE)

bam_config_file = system.file(package = "ssvQC", "extdata/ssvQC_bam_config.csv")
bam_config = QcConfigSignal.parse(bam_config_file)

sqc = ssvQC(feature_config, bam_config)

prepS
query_gr = feature_config$assessment_features$CTCF_features
sclust = ClusteredSignal_TSNE(sqc@signal_profile, sqc@feature_config$assessment_features)
sclust$
}
