% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcConfig.R
\name{.show_QcConfig}
\alias{.show_QcConfig}
\title{Title}
\usage{
.show_QcConfig(qc)
}
\arguments{
\item{qc}{}
}
\value{

}
\description{
Title
}
\examples{
feature_config_file = system.file(package = "ssvQC", "extdata/ssvQC_peak_config.csv")
config_df = .parse_config_body(feature_config_file)
qc = QcConfigFeatures(config_df)
}
