% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcConfigFeatures.R
\name{featuresList,QcConfigFeatures-method}
\alias{featuresList,QcConfigFeatures-method}
\title{Title}
\usage{
\S4method{featuresList}{QcConfigFeatures}(object)
}
\arguments{
\item{QcConfigFeatures}{}
}
\value{

}
\description{
Title
}
