% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcConfigFeatures.R
\name{featuresOverlaps,QcConfigFeatures-method}
\alias{featuresOverlaps,QcConfigFeatures-method}
\title{Title}
\usage{
\S4method{featuresOverlaps}{QcConfigFeatures}(object)
}
\arguments{
\item{QcConfigFeatures}{}
}
\value{

}
\description{
Title
}
