% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcConfig.R
\name{featuresQuery,QcConfigFeatures-method}
\alias{featuresQuery,QcConfigFeatures-method}
\title{Title}
\usage{
\S4method{featuresQuery}{QcConfigFeatures}(object)
}
\arguments{
\item{QcConfigFeatures}{}
}
\value{

}
\description{
Title
}
