% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcConfigSignal.R
\name{fetch_signal_at_features}
\alias{fetch_signal_at_features}
\title{Title}
\usage{
fetch_signal_at_features(qc_signal, query_gr)
}
\arguments{
\item{qc_signal}{}

\item{query_gr}{}
}
\value{

}
\description{
Title
}
\examples{
bam_config_file = system.file(package = "ssvQC", "extdata/ssvQC_bam_config.csv")
qc_signal = QcConfigSignal.parse(bam_config_file)

feature_config_file = system.file(package = "ssvQC", "extdata/ssvQC_peak_config.csv")
qc_features = QcConfigFeatures.parse(feature_config_file)
query_gr = qc_features$assessment_features
fetch_signal_at_features(qc_signal, query_gr)
}
