% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_misc.R
\name{guess_feature_file_format}
\alias{guess_feature_file_format}
\title{guess_feature_file_format}
\usage{
guess_feature_file_format(feature_files)
}
\arguments{
\item{feature_files}{A feature or interval file path. A vector of multiple files is acceptable.}
}
\value{
A character vector describing each entry in feature_files.
}
\description{
Check if feature file has a common format. ie. narrowPeak, broadPeak, or bed.
}
\examples{
feature_files = dir(system.file("extdata", package = "ssvQC"), pattern = "bed$|Peak$", full.names = TRUE)
guess_feature_file_format(feature_files)
}
