% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_scc.R
\name{harmonize_seqlengths}
\alias{harmonize_seqlengths}
\title{Forces seqlengths in GRanges to be that in header of bam_file}
\usage{
harmonize_seqlengths(gr, bam_file)
}
\arguments{
\item{gr}{GRanges}

\item{bam_file}{bam_file}
}
\value{
GRanges with seqlengths matching bam_file
}
\description{
This is important to avoid errors while using the gr to query the bam_file.
}
\examples{
peak_files = dir(system.file("extdata", package = "seqqc"), pattern = "Peak$", full.names = TRUE)
peak_grs = seqsetvis::easyLoad_narrowPeak(peak_files)
query_gr = resize(seqsetvis::ssvOverlapIntervalSets(peak_grs), 6e2, fix = "center")

bam_files = dir(system.file("extdata", package = "seqqc"), pattern = "^M.+bam$", full.names = TRUE)

query_gr.harmonized = seqqc:::harmonize_seqlengths(query_gr, bam_files[1])
seqlengths(query_gr.harmonized)

}
