% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_make.R
\name{make_anno_dt}
\alias{make_anno_dt}
\title{make_anno_dt}
\usage{
make_anno_dt(peak_grs, anno_grs, name_lev = NULL)
}
\arguments{
\item{peak_grs}{a named list of GRanges where each feature set should be
annotated. Use output from \link[seqsetvis]{easyLoad_narrowPeak} or
similar.}

\item{anno_grs}{a named list of GRanges where each corresponds to an
annotated feature type. Use output from \code{\link{make_anno_grs}}}

\item{name_lev}{Optional levels to impose order in feature sets.  Default of
NULL uses input order of peak_grs.}
}
\value{
data.table with annotation overlaps for inerval sets in peak_grs.
}
\description{
make_anno_dt
}
\examples{
gtf_file = system.file(package = "seqqc", "extdata/gencode.v35.annotation.at_peaks.gtf")
anno_grs = make_anno_grs(gtf_file)

peak_files = dir(system.file("extdata", package = "seqqc"),
  pattern = "Peak$", full.names = TRUE)
peak_grs = seqsetvis::easyLoad_narrowPeak(peak_files)

make_anno_dt(peak_grs, anno_grs)
}
