% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_make.R
\name{make_dt}
\alias{make_dt}
\title{make_dt}
\usage{
make_dt(files, group_lev = NULL, max_name_len = 30)
}
\arguments{
\item{files}{character vector of file paths.}

\item{group_lev}{If provided, should be directory names at equivalent depth
in file paths. Default of NULL disables this functionality and group will
be none and batch will be A.}

\item{max_name_len}{Names are truncated to this length. Default is 30.}
}
\value{
A data.table with attribute "file" and "name", "group", and "batch" attributes.
}
\description{
Creates a data.table from a vector of files.
}
\examples{
files = c("exp1/file1", "exp1/file2", "exp2/file3")
#default no group levels
make_dt(files)
dt1 = make_dt(files, group_lev = c("exp1", "exp2"))
levels(dt1$group)
dt1$batch
#reversed group_lev
dt2 = make_dt(files, group_lev = rev(c("exp1", "exp2")))
levels(dt2$group)
dt2$batch
}
