% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_make.R
\name{make_fq_dt}
\alias{make_fq_dt}
\title{make_fq_dt}
\usage{
make_fq_dt(
  fastq_files,
  fastq_names = basename(fastq_files),
  fastq_treatments = fastq_names,
  n_cores = getOption("mc.cores", 1),
  cache_counts = TRUE
)
}
\arguments{
\item{fastq_files}{paths to fast files (can be gzipped with .gz extension)}

\item{fastq_names}{optional parallel vector of names for fastq files.
Defaults to basename of fastq_files. Should be unique.}

\item{fastq_treatments}{optional parallel vector of treatments. Defaults to
fastq_names. May be duplicated.}

\item{n_cores}{number of cores to use to count lines in fastq files. Defaults
to mc.cores if set or 1.}

\item{cache_counts}{logical. Should the counts be saved to *.cnt files
alongside the fastq_files?  Default is TRUE}
}
\value{
a data.table countaining fastq, count, name, and treatment attributes
}
\description{
make_fq_dt
}
\examples{
fq_files = dir("inst/extdata", pattern = "(fq$)|(fq.gz$)|(fastq$)|(fastq.gz$)", full.names = TRUE)
 #no idea why this make_fq_dt example won't run
 #make_fq_dt(fq_files,
 #  fastq_names = c("4_reads_fq", "4_reads_gz", "5_reads_fq", "5_reads_gz"),
 #  cache_counts = FALSE)
}
