% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_make.R
\name{make_frip_dt}
\alias{make_frip_dt}
\title{make_frip_dt}
\usage{
make_frip_dt(
  query_dt,
  query_gr,
  n_cores = getOption("mc.cores", 1),
  name_lev = NULL
)
}
\arguments{
\item{query_dt}{data.table with query information. Only really needs file as
first column.}

\item{query_gr}{GRanges to be centered.}

\item{n_cores}{Number of cores to use. Defaults to mc.cores if set or 1.}

\item{name_lev}{Optional levels to impose order in feature sets.  Default of
NULL uses decreasing FRIP score.}
}
\value{
a data.table with reads_in_peak, mapped_reads, and frip data for all
  bam files in query_dt at each region in query_gr.
}
\description{
create a data.table with FRIP data.  To be used with
\code{\link{plot_frip_dt}}
}
\examples{
peak_file = dir(system.file("extdata", package = "seqqc"),
  pattern = "test_peaks.bed$", full.names = TRUE)
bam_file = dir(system.file("extdata", package = "seqqc"),
  pattern = "test_peaks.bam$", full.names = TRUE)

query_gr = seqsetvis::easyLoad_bed(peak_file)[[1]]

make_frip_dt(bam_file, query_gr)
}
