% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_make.R
\name{make_peak_dt}
\alias{make_peak_dt}
\title{make_peak_dt}
\usage{
make_peak_dt(peak_grs, treatments = NULL)
}
\arguments{
\item{peak_grs}{list of GRanges for peak sets}

\item{treatments}{opional character vector of treatments for each peak set.}
}
\value{
a data.table with peak_count data for each GRanges in peak_grs.
}
\description{
make_peak_dt
}
\examples{
peak_files = dir(system.file("extdata", package = "seqqc"),
  pattern = "Peak$", full.names = TRUE)
peak_grs = seqsetvis::easyLoad_narrowPeak(peak_files)

make_peak_dt(peak_grs)
}
