% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_make.R
\name{make_scc_dt.single}
\alias{make_scc_dt.single}
\title{make_scc_dt.single}
\usage{
make_scc_dt.single(
  bam_file,
  query_gr,
  frag_sizes,
  fetch_size = 3 * max(frag_sizes),
  cache_path = "~/.cache_peakrefine",
  cache_version = "v1",
  force_overwrite = FALSE,
  n_cores = getOption("mc.cores", 1L),
  ...
)
}
\arguments{
\item{bam_file}{a single path to a bam file}

\item{query_gr}{GRanges of regions to calculate SCC for}

\item{frag_sizes}{optional numeric. Fragment sizes to calculate correlation
at.  The higher the resolution the longer calculation will take.  The
default is to count by 10 from 50 to 350.}

\item{fetch_size}{optional numeric. Size in bp centered around each interval
in query_gr to retrieve.  Should be greater than max frag_size. The default
is 3*max(frag_sizes).}

\item{cache_path}{path to cache location for BiocFileCache to use.}

\item{cache_version}{Modifying the cache version will force recalulation of
all results going forward. Default is v1.}

\item{force_overwrite}{Logical, if TRUE, cache contents will be overwritten.}

\item{n_cores}{Number of cores to use. Defaults to mc.cores if set or 1.}

\item{...}{passed to Rsamtools::ScanBamParam()}
}
\value{
list fo tidy data.table of SCC data for bam_file
}
\description{
based on peakrefine::calcSCCMetrics
}
\examples{
peak_file = dir(system.file("extdata", package = "seqqc"),
  pattern = "test_peaks.bed$", full.names = TRUE)
bam_file = dir(system.file("extdata", package = "seqqc"),
  pattern = "test_peaks.bam$", full.names = TRUE)

query_gr = seqsetvis::easyLoad_bed(peak_file)[[1]]

scc_res = seqqc:::make_scc_dt.single(bam_file, query_gr, seq(50,300, by = 10))
}
