% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_misc.R
\name{parse_fetch_options}
\alias{parse_fetch_options}
\title{Title}
\usage{
parse_fetch_options(fop)
}
\arguments{
\item{fop}{}
}
\description{
Title
}
\examples{
fop = "win_size:10,win_method:\"summary\",summary_FUN:mean"
parse_fetch_options(fop)
}
