% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_anno_overlap}
\alias{plot_anno_overlap}
\title{plot_anno_overlap}
\usage{
plot_anno_overlap(anno_dt, name_lev = NULL)
}
\arguments{
\item{anno_dt}{Output from \code{\link{make_anno_dt}}}

\item{name_lev}{Optional name levels top apply.}
}
\value{
bar plot of annotation overlap frequency
}
\description{
plot_anno_overlap
}
\examples{
gtf_file = system.file(package = "seqqc", "extdata/gencode.v35.annotation.at_peaks.gtf")
anno_grs = make_anno_grs(gtf_file)

peak_files = dir(system.file("extdata", package = "seqqc"), pattern = "Peak$", full.names = TRUE)
peak_grs = seqsetvis::easyLoad_narrowPeak(peak_files)
names(peak_grs) = sub("_rand.+", "", names(peak_grs))

anno_dt = make_anno_dt(peak_grs, anno_grs)
plot_anno_overlap(anno_dt)
}
