% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_feature_overlap_signal_profiles}
\alias{plot_feature_overlap_signal_profiles}
\title{plot_feature_overlap_signal_profiles}
\usage{
plot_feature_overlap_signal_profiles(
  grouped_prof_dt,
  group_var = "overlap_group",
  rank_var = "rnk",
  fill_limits = NULL,
  signal_var = "y"
)
}
\arguments{
\item{grouped_prof_dt}{Output of \code{make_feature_overlap_signal_profiles}}

\item{group_var}{character. Must already be present in grouped_prof_dt.  Use \code{make_feature_overlap_signal_profiles} to add properly. Default is "overlap_group".}

\item{rank_var}{character. Must already be present in grouped_prof_dt.  Use \code{make_feature_overlap_signal_profiles} to add properly. Default is "rnk".}

\item{fill_limits}{Limits of heatmap fill scale.}

\item{signal_var}{value to plot}
}
\value{
list of ggplot plots that use feature overlaps to plot signal
}
\description{
plot_feature_overlap_signal_profiles
}
\examples{
bw_files = dir(system.file("extdata", package = "seqqc"), pattern = "^M.+bw$", full.names = TRUE)
query_dt = make_dt(bw_files)
query_dt[, sample := sub("_FE_random100.A", "", name)]

peak_files = dir(system.file("extdata", package = "seqqc"), pattern = "Peak$", full.names = TRUE)
names(peak_files) = sub("_CTCF_rand.+", "", basename(peak_files))
peak_grs = seqsetvis::easyLoad_narrowPeak(peak_files, )
overlaps_gr = seqsetvis::ssvOverlapIntervalSets(peak_grs)
query_gr = resize(overlaps_gr, 6e2, fix = "center")

group_prof_dt = make_feature_overlap_signal_profiles(query_dt, overlaps_gr)
plot_feature_overlap_signal_profiles(group_prof_dt)
}
