% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_fq_dt}
\alias{plot_fq_dt}
\title{plot_fq_dt}
\usage{
plot_fq_dt(fq_dt)
}
\arguments{
\item{fq_dt}{output of \code{\link{make_fq_dt}}}
}
\value{
a ggplot shoing read counts from fastq files
}
\description{
plot_fq_dt
}
\examples{
fq_files = dir("inst/extdata",
  pattern = "(fq$)|(fq.gz$)|(fastq$)|(fastq.gz$)",
  full.names = TRUE)
#no idea why this make_fq_dt example won't run
#fq_dt = make_fq_dt(fq_files,
#   fastq_names = c("4_reads_fq", "4_reads_gz", "5_reads_fq", "5_reads_gz"))
#plot_fq_dt(fq_dt)
}
