% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_scc_dt}
\alias{plot_scc_dt}
\title{plot_scc_dt}
\usage{
plot_scc_dt(scc_dt, main_title = NULL)
}
\arguments{
\item{scc_dt}{output from \code{\link{make_scc_dt}}}
}
\value{
list of ggplots relevant to SCC
}
\description{
plot strand cross correlation (SCC) info
}
\examples{
bam_files = dir(system.file("extdata", package = "seqqc"), pattern = "^M.+bam$", full.names = TRUE)
query_dt.bam = make_dt(bam_files)

peak_files = dir(system.file("extdata", package = "seqqc"), pattern = "Peak$", full.names = TRUE)
peak_grs = seqsetvis::easyLoad_narrowPeak(peak_files)
query_gr = resize(seqsetvis::ssvOverlapIntervalSets(peak_grs), 6e2, fix = "center")

scc_dt = make_scc_dt(query_dt.bam, query_gr)
plot_scc_dt(scc_dt)
}
