% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_signals}
\alias{plot_signals}
\title{plot_signals}
\usage{
plot_signals(
  prof_dt,
  query_gr,
  assign_dt = NULL,
  n_to_plot = 500,
  fill_var = "y_relative",
  anno_grs = NULL,
  frip_dt = NULL,
  scc_dt = NULL,
  name_lev = NULL,
  nclust = 6
)
}
\arguments{
\item{prof_dt}{data.table of signal profiles from
\link[seqsetvis]{ssvFetchBam} or \link[seqsetvis]{ssvFetchBigwig}.}

\item{query_gr}{The GRanges used to produce prof_dt.}

\item{assign_dt}{(optional) Precalculated clustering.  From running
\link[seqsetvis]{ssvSignalClustering} followed by
\code{\link{make_assign_dt}}}

\item{n_to_plot}{numeric. Number of items plotted in heatmap.  Default is 500.}

\item{fill_var}{character. Variable in prof_dt to cluster and plot on. Default
of y_relative is the fraction of max y per id.}

\item{anno_grs}{(optional) output from \code{\link{make_anno_grs}}}

\item{frip_dt}{(optional) output from \code{\link{make_frip_dt}}}

\item{scc_dt}{(optional) output from \code{\link{make_scc_dt}}}

\item{name_lev}{(optional) manual specification for levels of "name" to
control ordering.}

\item{nclust}{Number of clusters for heatmap. Defaults to 6.}
}
\value{
named list of ggplot plots and relevant data
}
\description{
Performs clustering on signal profiles in prof_dt and produces various plots.
Optionally outputs annotation (with anno_grs) and/or FRIP (with frip_dt) plots
that incorporate signal clustering information.
}
\examples{
bw_files = dir(system.file("extdata", package = "seqqc"), pattern = "^M.+bw$", full.names = TRUE)
query_dt = make_dt(bw_files)
query_dt[, sample := sub("_FE_random100.A", "", name)]

peak_files = dir(system.file("extdata", package = "seqqc"), pattern = "Peak$", full.names = TRUE)
peak_grs = seqsetvis::easyLoad_narrowPeak(peak_files)
query_gr = resize(seqsetvis::ssvOverlapIntervalSets(peak_grs), 6e2, fix = "center")

prof_dt = seqsetvis::ssvFetchBigwig(query_dt, query_gr, return_data.table = TRUE)

sig_res = plot_signals(prof_dt, query_gr)
sig_res$heatmap
sig_res$heatmap_sidebar
sig_res$cluster_assignment

gtf_file = system.file(package = "seqqc", "extdata/gencode.v35.annotation.at_peaks.gtf")
anno_grs = make_anno_grs(gtf_file)

sig_res.anno = plot_signals(prof_dt, query_gr, anno_grs = anno_grs)
sig_res.anno$heatmap
sig_res.anno$annotation_heatmap

bam_files = dir(system.file("extdata", package = "seqqc"), pattern = "^M.+bam$", full.names = TRUE)
query_dt.bam = make_dt(bam_files)

frip_dt = make_frip_dt(query_dt.bam, query_gr)

sig_res.frip = plot_signals(prof_dt, query_gr, frip_dt = frip_dt)
sig_res.frip$heatmap
sig_res.frip$frip_bars_per_cluster
sig_res.frip$frip_text_per_cluster

scc_dt = make_scc_dt(query_dt.bam, query_gr)
sig_res.scc = plot_signals(prof_dt, query_gr, scc_dt = scc_dt)

sig_res.scc$scc_curves_per_cluster
sig_res.scc$scc_dots_per_cluster

}
