% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tsne_plots_glyphs.R
\name{plot_summary_glyph}
\alias{plot_summary_glyph}
\title{Title}
\usage{
plot_summary_glyph(
  summary_dt,
  x_points,
  y_points = x_points,
  xrng = c(-0.5, 0.5),
  yrng = c(-0.5, 0.5),
  ylim = NULL,
  p = NULL,
  N_floor = 0,
  N_ceiling = NULL,
  min_size = 0.3,
  return_data = FALSE,
  color_mapping = NULL,
  x_var = "x",
  y_var = "y",
  wide_var = "name"
)
}
\arguments{
\item{summary_dt}{}

\item{x_points}{}

\item{y_points}{}

\item{xrng}{}

\item{yrng}{}

\item{ylim}{}

\item{p}{}

\item{N_floor}{}

\item{N_ceiling}{}

\item{min_size}{}

\item{return_data}{}

\item{color_mapping}{}

\item{x_var}{}

\item{y_var}{}

\item{wide_var}{}
}
\description{
Title
}
