% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcConfig.R
\name{QcColorMapping}
\alias{QcColorMapping}
\alias{QcColorMapping,QcConfig-method}
\title{QcColorMapping}
\usage{
QcColorMapping(object)

\S4method{QcColorMapping}{QcConfig}(object)
}
\arguments{
\item{object}{a QcConfig object}
}
\value{
a named vector of colors
}
\description{
QcColorMapping

QcColorMapping for QcConfig
}
\examples{
QcColorMapping(QcConfig(c("A", "B")))
}
