% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcConfig.R
\docType{class}
\name{QcConfig-class}
\alias{QcConfig-class}
\title{QcConfig}
\description{
QcConfig
}
\section{Slots}{

\describe{
\item{\code{file_paths}}{character.}

\item{\code{groups}}{numeric.}

\item{\code{group_names}}{character.}

\item{\code{group_colors}}{character.}
}}

