% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcConfig.R
\docType{class}
\name{QcConfig-class}
\alias{QcConfig-class}
\alias{QcConfig.files}
\alias{QcColorMapping}
\alias{QcColorMapping,QcConfig-method}
\alias{QcScaleColor,QcConfig-method}
\alias{QcScaleFill}
\alias{QcScaleFill,QcConfig-method}
\alias{show,QcConfig-method}
\alias{QcConfig.save_config}
\title{QcConfig}
\usage{
QcConfig.files(
  file_paths,
  groups = NULL,
  group_names = NULL,
  group_colors = NULL
)

QcColorMapping(object)

\S4method{QcColorMapping}{QcConfig}(object)

\S4method{QcScaleColor}{QcConfig}(object)

QcScaleFill(object)

\S4method{QcScaleFill}{QcConfig}(object)

\S4method{show}{QcConfig}(object)

QcConfig.save_config(object, file)
}
\arguments{
\item{file_paths}{character paths to files}

\item{groups}{numeric vector of group assignments. 1 is first item in group_names, 2 is second, etc. Default is seq_along(file_path)}

\item{group_names}{vector of group names to assign from according to groups}

\item{group_colors}{vector of colors to use per group}

\item{object}{QcConfig object to save to text file.}

\item{QcConfig}{}
}
\value{
A QcConfig object

a named vector of colors

a ggplot2 scale_fill_manual

invisibly returns path to saved config file.
}
\description{
QcConfig

QcConfig

QcColorMapping

QcColorMapping for QcConfig

QcScaleColor for QcConfig

QcScaleFill

QcScaleFill for QcConfig

QcConfig.save_config
}
\section{Slots}{

\describe{
\item{\code{file_paths}}{character.}

\item{\code{groups}}{numeric.}

\item{\code{group_names}}{character.}

\item{\code{group_colors}}{character.}
}}

\examples{
QcConfig.files(c("A", "B"))
QcColorMapping(QcConfig(c("A", "B")))
my_df = data.frame(group = c("A", "B"))
my_df$x = 1:2
my_df$y = 1
library(ggplot2)
ggplot(my_df, aes(x = x, y = y, color = group)) +
  geom_point(size = 20) +
  expand_limits(x = c(0, 3)) +
  QcScaleColor(QcConfig(my_df$group))
my_df = data.frame(group = c("A", "B"))
my_df$x = 1:2
my_df$y = 1
library(ggplot2)
ggplot(my_df, aes(x = x, y = y, fill = group)) +
  geom_point(size = 20, pch = 22) +
  expand_limits(x = c(0, 3)) +
  QcScaleFill(QcConfig(my_df$group))
cfg_file = system.file("extdata/ssvQC_peak_config.csv", package = "ssvQC")
qc_features = QcConfigFeatures.parse(cfg_file)
f = tempfile()
QcConfig.save_config(qc_features, f)
}
