% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcConfig.R
\docType{class}
\name{QcConfigFeatures-class}
\alias{QcConfigFeatures-class}
\title{QcConfigFeatures}
\description{
QcConfigFeatures
}
\section{Slots}{

\describe{
\item{\code{feature_load_FUN}}{function.}

\item{\code{n_peaks}}{numeric.}

\item{\code{min_fraction}}{numeric.}

\item{\code{min_number}}{numeric.}
}}

