% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcConfigFeatures.R
\name{QcConfigFeatures.GRanges}
\alias{QcConfigFeatures.GRanges}
\title{Title}
\usage{
QcConfigFeatures.GRanges(
  query_gr,
  n_peaks = 2000,
  run_separately = TRUE,
  sample_names = NULL,
  sample_names.split = NULL,
  group_names = NULL,
  group_colors = NULL,
  balance_groups = FALSE,
  overlap_extension = 0,
  consensus_fraction = getOption("SQC_CONSENSUS_FRACTION", 0),
  consensus_n = getOption("SQC_CONSENSUS_N", 1),
  process_features = getOption("SQC_PROCESS_FEATURES", TRUE)
)
}
\arguments{
\item{query_gr}{either a list of GRanges of a GRanges object.  GRanges object will be handled differently if its mcols are a membership table as generated by seqsetvis::sssvOverlapIntervalSets or seqsetvis::ssvConsensusIntervalSets.}

\item{n_peaks}{number of peaks to subset for}

\item{run_separately}{If TRUE, each item is considered a separate group. Default is TRUE.}

\item{sample_names}{Unique names to use per GRanges in list or columns in membership table.  Do not set if single GRanges does not have memb table.}

\item{sample_names.split}{Unique names to use per GRanges in list or columns in membership table.  Do not set if single GRanges does not have memb table.}

\item{group_names}{vector of group names to assign from according to groups}

\item{group_colors}{vector of colors to use per group}

\item{balance_groups}{If TRUE, will attempt to represent imbalanced groups more equally}

\item{overlap_extension}{bp to extends regions by before calculating overlaps}

\item{consensus_fraction}{A numeric between 0 and 1 specifying the fraction of grs that must overlap to be considered consensus.}

\item{consensus_n}{An integer number specifying the absloute minimum of input grs that must overlap for a site to be considered consensus.}

\item{process_features}{if TRUE, features are loaded and overlapped to generate assement regions as part of creating QcConfigFeatures object}
}
\description{
Title
}
