% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcConfigFeatures.R
\name{QcConfigFeatures.null}
\alias{QcConfigFeatures.null}
\title{QcConfigFeatures null placeholder}
\usage{
QcConfigFeatures.null()
}
\value{
QcConfigFeatures object
}
\description{
QcConfigFeatures null placeholder
}
\examples{
QcConfigFeatures.null()
}
