% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcConfigFeatures.R
\name{QcConfigFeatures.overlap_run_by}
\alias{QcConfigFeatures.overlap_run_by}
\title{Title}
\usage{
QcConfigFeatures.overlap_run_by(
  object,
  group_colors = NULL,
  n_peaks = 1000,
  balance_groups = FALSE,
  overlap_extension = 0,
  consensus_fraction = getOption("SQC_CONSENSUS_FRACTION", 0),
  consensus_n = getOption("SQC_CONSENSUS_N", 1),
  process_features = getOption("SQC_PROCESS_FEATURES", TRUE)
)
}
\arguments{
\item{object}{}

\item{group_colors}{}

\item{n_peaks}{}

\item{balance_groups}{}

\item{overlap_extension}{}

\item{consensus_fraction}{}

\item{consensus_n}{}

\item{process_features}{}
}
\value{

}
\description{
Title
}
\examples{
np_files = dir(system.file(package = "ssvQC", "extdata"), pattern = "Peak$", full.names = TRUE)
object = QcConfigFeatures.files(np_files, run_separately = TRUE)
object = ssvQC.prepFeatures(object)
plot(object)
}
