% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcConfig.R
\name{QcConfigFeatures.parse}
\alias{QcConfigFeatures.parse}
\title{QcConfigFeatures.parse}
\usage{
QcConfigFeatures.parse(
  feature_config_file,
  process_features = getOption("SQC_PROCESS_FEATURES", TRUE)
)
}
\arguments{
\item{feature_config_file}{}
}
\value{

}
\description{
QcConfigFeatures.parse
}
\examples{
feature_config_file = system.file(package = "ssvQC", "extdata/ssvQC_peak_config.csv")
QcConfigFeatures.parse(feature_config_file)
}
