% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcConfig.R
\name{QcConfigSignal.parse}
\alias{QcConfigSignal.parse}
\title{Title}
\usage{
QcConfigSignal.parse(signal_config_file)
}
\arguments{
\item{signal_config_file}{}
}
\value{

}
\description{
Title
}
\examples{
bam_config_file = system.file(package = "ssvQC", "extdata/ssvQC_bam_config.csv")
QcConfigSignal.parse(bam_config_file)

bigwig_config_file = system.file(package = "ssvQC", "extdata/ssvQC_bigwig_config.csv")
QcConfigSignal.parse(bigwig_config_file)
}
