% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcConfig.R
\name{QcConfigSignal}
\alias{QcConfigSignal}
\title{Title}
\usage{
QcConfigSignal(
  config_df,
  run_by = "All",
  to_run = NULL,
  to_run_reference = NULL,
  color_by = "file",
  color_mapping = NULL,
  read_mode = NULL,
  view_size = getOption("SQC_VIEW_SIZE", 3000)
)
}
\arguments{
\item{config_df}{}

\item{run_by}{}

\item{color_by}{}

\item{color_mapping}{}

\item{read_mode}{}

\item{view_size}{}
}
\value{

}
\description{
Title
}
\examples{
bam_config_file = system.file(package = "ssvQC", "extdata/ssvQC_bam_config.csv")
bam_config_df = .parse_config_body(bam_config_file)
QcConfigSignal(bam_config_df)

bigwig_config_file = system.file(package = "ssvQC", "extdata/ssvQC_bigwig_config.csv")
bigwig_config_df = .parse_config_body(bigwig_config_file)
QcConfigSignal(bigwig_config_df)
}
