% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tsne.R
\name{run_tsne}
\alias{run_tsne}
\title{run_tsne}
\usage{
run_tsne(
  profile_dt,
  perplexity = 100,
  n_cores = getOption("mc.cores", 1),
  high_topright = TRUE,
  norm1 = TRUE,
  Y_init = NULL,
  verbose = TRUE,
  force_overwrite = FALSE,
  x_var = "x",
  y_var = "y",
  id_var = "id",
  wide_var = "name",
  tall_var = "tall_none"
)
}
\arguments{
\item{profile_dt}{}

\item{perplexity}{}

\item{n_cores}{}

\item{high_topright}{}

\item{norm1}{}

\item{Y_init}{}

\item{verbose}{}

\item{force_overwrite}{}

\item{wide_var}{}

\item{tall_var}{}
}
\description{
run_tsne
}
