% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_misc.R
\name{ssv_mclapply}
\alias{ssv_mclapply}
\title{ssv_mclapply}
\usage{
ssv_mclapply(X, FUN, mc.cores = getOption("mc.cores", 1), ...)
}
\value{
result of either pblapply or pbmclapply
}
\description{
ssv_mclapply
}
