% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssvQC.R
\name{ssvQC.plotMappedReads}
\alias{ssvQC.plotMappedReads}
\title{ssvQC.plotMappedReads}
\usage{
ssvQC.plotMappedReads(object, out_dir)
}
\arguments{
\item{object}{}
}
\value{

}
\description{
ssvQC.plotMappedReads
}
