% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssvQC_select.R
\name{ssvQC.removeClusters}
\alias{ssvQC.removeClusters}
\title{ssvQC.removeClusters}
\usage{
ssvQC.removeClusters(
  sqc,
  cluster_numbers,
  features_name = NULL,
  signals_name = NULL,
  invert = FALSE
)
}
\arguments{
\item{sqc}{}

\item{cluster_numbers}{}

\item{features_name}{}

\item{signals_name}{}

\item{invert}{}
}
\value{

}
\description{
ssvQC.removeClusters
}
\examples{
library(ssvQC)
options(mc.cores = 1)
set.seed(0)
features_config_file = system.file(package = "ssvQC", "extdata/ssvQC_peak_config.csv")
features_config = QcConfigFeatures.parse(features_config_file)

bam_config_file = system.file(package = "ssvQC", "extdata/ssvQC_bam_config.csv")
bam_config = QcConfigSignal.parse(bam_config_file)

sqc.complete.file = ssvQC(features_config_file, bam_config_file)

sqc.complete = ssvQC(features_config, bam_config)
sqc = ssvQC.runAll(sqc.complete)
sqc$plots$signal$heatmaps$CTCF_features$CTCF_signal

sqc.rmclust = ssvQC.removeClusters(sqc, cluster_numbers = 1:4)
sqc.rmclust$plots$signal$heatmaps$CTCF_features$CTCF_signal

sqc.rmclust_inv = ssvQC.removeClusters(sqc, cluster_numbers = 1:4, invert = TRUE)
sqc.rmclust_inv$plots$signal$heatmaps$CTCF_features$CTCF_signal
}
