% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssvQC.R
\name{ssvQC.runAll}
\alias{ssvQC.runAll}
\title{Title}
\usage{
ssvQC.runAll(sqc)
}
\arguments{
\item{sqc}{}
}
\value{

}
\description{
Title
}
\examples{
feature_config_file = system.file(package = "ssvQC", "extdata/ssvQC_peak_config.csv")
bam_config_file = system.file(package = "ssvQC", "extdata/ssvQC_bam_config.csv")

sqc = ssvQC(feature_config_file, bam_config_file)
}
