% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssvQC.R
\name{ssvQC.save_config}
\alias{ssvQC.save_config}
\title{ssvQC.save_config}
\usage{
ssvQC.save_config(object, file)
}
\arguments{
\item{object}{A ssvQC object}

\item{file}{A .features.csv and .signal.csv file will be written with this prefix.}
}
\value{
invisibly returns config file paths.
}
\description{
ssvQC.save_config
}
