% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssvQC_select.R
\name{ssvQC.removeClusters}
\alias{ssvQC.removeClusters}
\alias{ssvQC.selectClusters}
\title{ssvQC.removeClusters}
\usage{
ssvQC.removeClusters(
  sqc,
  cluster_numbers,
  features_name = NULL,
  signals_name = NULL,
  invert = FALSE
)

ssvQC.selectClusters(
  sqc,
  cluster_numbers,
  features_name = NULL,
  signals_name = NULL,
  invert = FALSE
)
}
\arguments{
\item{sqc}{A valid ssvQC object after clustering by ssvQC.prepSignal}

\item{cluster_numbers}{Number for clusters.}

\item{features_name}{feature_name to perform selection on. Required if multiple feature sets present.}

\item{signals_name}{signal_name to perform selection on. Required if multiple signal sets present.}

\item{invert}{If true, flip removal to selection and vice versa.}
}
\value{
ssvQC object with selected clusters removed.
}
\description{
ssvQC.removeClusters

ssvQC.selectClusters
}
\examples{
library(ssvQC)
options(mc.cores = 1)
set.seed(0)
features_config_file = system.file(package = "ssvQC", "extdata/ssvQC_peak_config.csv")
features_config = QcConfigFeatures.parse(features_config_file)

bam_config_file = system.file(package = "ssvQC", "extdata/ssvQC_bam_config.csv")
bam_config = QcConfigSignal.parse(bam_config_file)

sqc.complete.file = ssvQC(features_config_file, bam_config_file)

sqc.complete = ssvQC(features_config, bam_config)
sqc = ssvQC.runAll(sqc.complete)
sqc$plots$signal$heatmaps$CTCF_features$CTCF_signal

sqc.rmclust = ssvQC.removeClusters(sqc, cluster_numbers = 1:4)
sqc.rmclust$plots$signal$heatmaps$CTCF_features$CTCF_signal

sqc.rmclust_inv = ssvQC.removeClusters(sqc, cluster_numbers = 1:4, invert = TRUE)
sqc.rmclust_inv$plots$signal$heatmaps$CTCF_features$CTCF_signal

sqc.selclust = ssvQC.selectClusters(sqc, cluster_numbers = 1:4)
sqc.selclust$plots$signal$heatmaps$CTCF_features$CTCF_signal

sqc.selclust_inv = ssvQC.selectClusters(sqc, cluster_numbers = 1:4, invert = TRUE)
sqc.selclust_inv$plots$signal$heatmaps$CTCF_features$CTCF_signal
}
