% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_write.R
\name{write_bed_overlaps}
\alias{write_bed_overlaps}
\title{write_bed_overlaps}
\usage{
write_bed_overlaps(overlaps_gr, assign_dt, file = "regions_with_overlaps.txt")
}
\arguments{
\item{overlaps_gr}{GRanges of regions used to make assign_dt with memb table in mcols.}

\item{assign_dt}{data.table with "id" and "cluster_id" that maps region ids to clusters. In output from \code{\link{plot_signals}}.}

\item{file}{File to write to}
}
\value{
invisibly returns data that is written to file
}
\description{
write_bed_overlaps
}
\examples{
bw_files = dir(system.file("extdata", package = "seqqc"), pattern = "^M.+bw$", full.names = TRUE)
query_dt = make_dt(bw_files)
query_dt[, sample := sub("_FE_random100.A", "", name)]

peak_files = dir(system.file("extdata", package = "seqqc"), pattern = "Peak$", full.names = TRUE)
peak_grs = seqsetvis::easyLoad_narrowPeak(peak_files)
overlaps_gr = seqsetvis::ssvOverlapIntervalSets(peak_grs)
query_gr = resize(overlaps_gr, 6e2, fix = "center")

prof_dt = seqsetvis::ssvFetchBigwig(query_dt, query_gr, return_data.table = TRUE)

sig_res = plot_signals(prof_dt, query_gr)
assign_dt = sig_res$cluster_assignment

outdir = system.file("extdata", package = "seqqc")
write_bed_overlaps(overlaps_gr, assign_dt,
  file = file.path(outdir, "regions_with_overlaps.txt"))
}
