% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_ssvQC.per_peak}
\alias{write_ssvQC.per_peak}
\title{write_ssvQC.per_peak}
\usage{
write_ssvQC.per_peak(sqc, out_dir = get_wd())
}
\arguments{
\item{sqc}{}

\item{f}{}
}
\value{

}
\description{
write_ssvQC.per_peak
}
