% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_ssvQC.per_peak}
\alias{write_ssvQC.per_peak}
\alias{write_ssvQC.summary}
\alias{write_ssvQC.correlation}
\title{write_ssvQC.per_peak}
\usage{
write_ssvQC.per_peak(sqc, out_dir = getwd())

write_ssvQC.summary(sqc, out_dir = getwd())

write_ssvQC.correlation(sqc, out_dir = getwd())
}
\arguments{
\item{sqc}{A valid ssvQC object, FRIP and SCC must be present.}

\item{out_dir}{Output directory path. Defaults to current directory.}
}
\value{
invisibly returns paths to per peak QC files.
}
\description{
write_ssvQC.per_peak

write_ssvQC.summary
}
\examples{
set.seed(0)
features_config_file = system.file(
  package = "ssvQC", 
  "extdata/ssvQC_peak_config.csv"
)
features_config = QcConfigFeatures.parse(features_config_file)

bam_config_file = system.file(
  package = "ssvQC", 
  "extdata/ssvQC_bam_config.csv"
)
bam_config = QcConfigSignal.parse(bam_config_file)

sqc = ssvQC(features_config, bam_config)
sqc = ssvQC.runAll(sqc)
out = tempdir()
qc_files.per_peak = write_ssvQC.per_peak(sqc, out)
qc_files.summary = write_ssvQC.summary(sqc, out)
qc_files.correlation = write_ssvQC.correlation(sqc, out)
dir(out)
}
