% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_scc.R
\name{bfcif}
\alias{bfcif}
\title{Title}
\usage{
bfcif(
  bfc,
  rname,
  FUN,
  version = getOption("SQC_CACHE_VERSION", "v3"),
  force_overwrite = getOption("SQC_FORCE_CACHE_OVERWRITE", FALSE)
)
}
\arguments{
\item{bfc}{}

\item{rname}{}

\item{FUN}{}

\item{version}{}

\item{force_overwrite}{}
}
\description{
Title
}
