% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteredSignal.R
\docType{class}
\name{ClusteredSignal-class}
\alias{ClusteredSignal-class}
\title{ClusteredSignal}
\description{
ClusteredSignal
}
\section{Slots}{

\describe{
\item{\code{signal_data}}{}

\item{\code{query_gr}}{}

\item{\code{signal_var}}{}

\item{\code{facet_var}}{}

\item{\code{extra_var}}{}
}}

