% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteredSignal.R
\name{ClusteredSignal.fromConfig}
\alias{ClusteredSignal.fromConfig}
\title{ClusteredSignal.fromConfig}
\usage{
ClusteredSignal.fromConfig(
  signal_config,
  query_gr,
  manual_assigned = list(),
  nclust = 6,
  facet_var = "name_split",
  extra_var = character(),
  bfc = new_cache()
)
}
\arguments{
\item{signal_config}{A valid QcConfigSignal}

\item{query_gr}{A GRanges containing regions to retrieve signal data at.}

\item{manual_assigned}{NYI but should allow manual cluster assignment.}

\item{nclust}{Number of k-means clusters to calculate. Default is 6.}

\item{facet_var}{Variable that will eventually be used in heatmap facets.  Ensures it is preserved and not aggregated away.  Default is "name_split".}

\item{extra_var}{Any extra variables to preserve and avaoid aggregating away.}

\item{bfc}{BiocFileCache to use, uses default location otherwise.}
}
\value{
A ClusteredSignal object containing clustered signal profiles.
}
\description{
ClusteredSignal.fromConfig
}
\examples{
options(mc.cores = 10)
set.seed(0)
feature_config_file = system.file(package = "ssvQC", "extdata/ssvQC_peak_config.csv")
feature_config = QcConfigFeatures.parse(feature_config_file, process_features= TRUE)

bam_config_file = system.file(package = "ssvQC", "extdata/ssvQC_bam_config.csv")
bam_config = QcConfigSignal.parse(bam_config_file)

sqc = ssvQC(feature_config, bam_config)

prepS
query_gr = feature_config$assessment_features$CTCF_features
sclust = ClusteredSignal(sqc@signal_profile, sqc@feature_config$assessment_features)
sclust$
}
