% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteredSignal.R
\name{ClusteredSignal}
\alias{ClusteredSignal}
\title{ClusteredSignal}
\usage{
ClusteredSignal(
  signal_profile_dt,
  query_gr,
  manual_assigned = list(),
  nclust = 6,
  signal_var = "y",
  signal_var.within = "y",
  facet_var = "name_split",
  extra_var = character(),
  bfc = new_cache()
)
}
\arguments{
\item{signal_profile_dt}{}

\item{query_gr}{}

\item{manual_assigned}{}

\item{nclust}{}

\item{signal_var}{}

\item{signal_var.within}{}

\item{facet_var}{}

\item{extra_var}{}

\item{bfc}{BiocFileCache to use, uses default location otherwise.}
}
\description{
ClusteredSignal
}
