% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteredSignal.R
\name{ClusteredSignal}
\alias{ClusteredSignal}
\title{ClusteredSignal}
\usage{
ClusteredSignal(
  signal_data,
  query_gr,
  manual_assigned = list(),
  nclust = 6,
  signal_var = "y",
  facet_var = "name_split",
  extra_var = character()
)
}
\value{

}
\description{
ClusteredSignal
}
\examples{
signal_data = sqc@signal_profile
query_gr = sqc@feature_config$assessment_features

}
