% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcConfig.R
\name{.show_QcConfig}
\alias{.show_QcConfig}
\title{.show_QcConfig}
\usage{
.show_QcConfig(qc)
}
\arguments{
\item{qc}{A QcConfig object}
}
\description{
used by show
}
\examples{
feature_config_file = system.file(package = "ssvQC", "extdata/ssvQC_peak_config.csv")
qc = QcConfigFeatures.parse(feature_config_file)
qc
}
