% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_misc.R
\name{get_args}
\alias{get_args}
\title{get_args}
\usage{
get_args(env = parent.frame(), to_ignore = character(), ...)
}
\arguments{
\item{env}{}

\item{...}{}
}
\description{
returns parameters of calling function as a named list.
}
