% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_misc.R
\name{get_feature_file_load_function}
\alias{get_feature_file_load_function}
\title{Title}
\usage{
get_feature_file_load_function(feature_files)
}
\arguments{
\item{feature_files}{}
}
\description{
Title
}
