% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssvQC.R
\name{guess_feature_file_format}
\alias{guess_feature_file_format}
\title{guess_feature_file_format}
\usage{
guess_feature_file_format(feature_files)
}
\arguments{
\item{feature_files}{}
}
\value{

}
\description{
guess_feature_file_format
}
