% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcConfig.R
\name{QcConfig}
\alias{QcConfig}
\title{QcConfig}
\usage{
QcConfig(file_paths, groups = NULL, group_names = NULL, group_colors = NULL)
}
\arguments{
\item{file_paths}{character paths to files}

\item{groups}{numeric vector of group assignments. 1 is first item in group_names, 2 is second, etc. Default is seq_along(file_path)}

\item{group_names}{vector of group names to assign from according to groups}

\item{group_colors}{vector of colors to use per group}
}
\value{
A QcConfig object
}
\description{
QcConfig
}
\examples{
QcConfig(c("A", "B"))
}
