% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcConfigFeatures.R
\name{QcConfigFeatures.overlap_run_by}
\alias{QcConfigFeatures.overlap_run_by}
\title{Title}
\usage{
QcConfigFeatures.overlap_run_by(
  object,
  group_colors = NULL,
  n_peaks = 1000,
  balance_groups = FALSE,
  overlap_extension = 0,
  consensus_fraction = getOption("SQC_CONSENSUS_FRACTION", 0),
  consensus_n = getOption("SQC_CONSENSUS_N", 1),
  process_features = getOption("SQC_PROCESS_FEATURES", TRUE)
)
}
\arguments{
\item{object}{}

\item{group_colors}{}

\item{n_peaks}{}

\item{balance_groups}{If TRUE, will attempt to represent imbalanced groups more equally}

\item{overlap_extension}{}

\item{consensus_fraction}{A numeric between 0 and 1 specifying the fraction of grs that must overlap to be considered consensus.}

\item{consensus_n}{An integer number specifying the absloute minimum of input grs that must overlap for a site to be considered consensus.}

\item{process_features}{if TRUE, features are loaded and overlapped to generate assement regions as part of creating QcConfigFeatures object}
}
\description{
Title
}
\examples{
np_files = dir(system.file(package = "ssvQC", "extdata"), pattern = "Peak$", full.names = TRUE)
object = QcConfigFeatures.files(np_files, run_separately = TRUE)
object = ssvQC.prepFeatures(object)
plot(object)
}
