% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcConfig.R
\name{QcConfigFeatures}
\alias{QcConfigFeatures}
\title{QcConfigFeatures}
\usage{
QcConfigFeatures(
  cfg_df,
  group_var = "All",
  group_to_run = NULL,
  group_reference = NULL,
  color_var = "file",
  color_mapping = NULL,
  feature_load_FUN = NULL,
  n_peaks = 1000,
  min_fraction = DEFAULT_CONSENSUS_FRACTION,
  min_number = DEFAULT_CONSENSUS_N
)
}
\arguments{
\item{cfg_df}{}

\item{group_var}{}

\item{color_var}{}

\item{color_mapping}{}

\item{feature_load_FUN}{}

\item{n_peaks}{}

\item{min_fraction}{}

\item{min_number}{}
}
\value{

}
\description{
QcConfigFeatures
}
\examples{
feature_config_file = system.file(package = "ssvQC", "extdata/ssvQC_peak_config.csv")
cfg_df = .parse_config_body(feature_config_file)
QcConfigFeatures(cfg_df)
}
