% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcConfigFeatures.R
\name{QcConfigFeatures}
\alias{QcConfigFeatures}
\title{QcConfigFeatures}
\usage{
QcConfigFeatures(
  config_df,
  run_by = "All",
  to_run = NULL,
  to_run_reference = NULL,
  color_by = "file",
  color_mapping = NULL,
  feature_load_FUN = NULL,
  n_peaks = 1000,
  overlap_extension = 0,
  consensus_fraction = getOption("SQC_CONSENSUS_FRACTION", 0),
  consensus_n = getOption("SQC_CONSENSUS_N", 1),
  process_features = getOption("SQC_PROCESS_FEATURES", TRUE)
)
}
\arguments{
\item{config_df}{}

\item{run_by}{}

\item{color_by}{}

\item{color_mapping}{}

\item{feature_load_FUN}{}

\item{n_peaks}{}

\item{consensus_fraction}{}

\item{consensus_n}{}
}
\value{

}
\description{
QcConfigFeatures
}
\examples{
feature_config_file = system.file(package = "ssvQC", "extdata/ssvQC_peak_config.csv")
config_df = .parse_config_body(feature_config_file)
config_df$file = file.path(system.file(package = "ssvQC", "extdata"), config_df$file)
feature_conf = QcConfigFeatures(config_df, process_features = TRUE)
}
