% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcConfig.R
\name{QcConfigFeatures}
\alias{QcConfigFeatures}
\title{QcConfigFeatures}
\usage{
QcConfigFeatures(
  config_df,
  run_by = "All",
  to_run = NULL,
  to_run_reference = NULL,
  color_by = "file",
  color_mapping = NULL,
  feature_load_FUN = NULL,
  n_peaks = 1000,
  overlap_extension = 0,
  consensus_fraction = DEFAULT_CONSENSUS_FRACTION,
  consensus_n = DEFAULT_CONSENSUS_N
)
}
\arguments{
\item{config_df}{}

\item{run_by}{}

\item{color_by}{}

\item{color_mapping}{}

\item{feature_load_FUN}{}

\item{n_peaks}{}

\item{consensus_fraction}{}

\item{consensus_n}{}
}
\value{

}
\description{
QcConfigFeatures
}
\examples{
feature_config_file = system.file(package = "ssvQC", "extdata/ssvQC_peak_config.csv")
config_df = .parse_config_body(feature_config_file)
QcConfigFeatures(config_df)
}
