% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcConfig.R
\name{QcConfigFeatures}
\alias{QcConfigFeatures}
\title{QcConfigFeatures}
\usage{
QcConfigFeatures(
  file_paths,
  groups = NULL,
  group_names = NULL,
  group_colors = NULL,
  n_peaks = 5000,
  min_fraction = 0,
  min_number = 1
)
}
\arguments{
\item{file_paths}{character paths to files}

\item{groups}{numeric vector of group assignments. 1 is first item in group_names, 2 is second, etc. Default is seq_along(file_path)}

\item{group_names}{vector of group names to assign from according to groups}

\item{group_colors}{vector of colors to use per group}

\item{n_peaks}{number of peaks to subset for}

\item{min_fraction}{A numeric between 0 and 1 specifying the fraction of grs that must overlap to be considered consensus.}

\item{min_number}{An integer number specifying the absloute minimum of input grs that must overlap for a site to be considered consensus.}
}
\value{
a QcConfigFeatures object
}
\description{
QcConfigFeatures
}
\examples{
QcConfigFeatures(c("A", "B"))
}
