% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcConfig.R
\name{QcConfigSignal}
\alias{QcConfigSignal}
\title{Title}
\usage{
QcConfigSignal(
  cfg_df,
  group_var = "All",
  group_to_run = NULL,
  group_reference = NULL,
  color_var = "file",
  color_mapping = NULL,
  read_mode = NULL,
  view_size = DEFAULT_VIEW_SIZE
)
}
\arguments{
\item{cfg_df}{}

\item{group_var}{}

\item{color_var}{}

\item{color_mapping}{}

\item{read_mode}{}

\item{view_size}{}
}
\value{

}
\description{
Title
}
\examples{
bam_config_file = system.file(package = "ssvQC", "extdata/ssvQC_bam_config.csv")
bam_cfg_df = .parse_config_body(bam_config_file)
QcConfigSignal(bam_cfg_df)

bigwig_config_file = system.file(package = "ssvQC", "extdata/ssvQC_bigwig_config.csv")
bigwig_cfg_df = .parse_config_body(bigwig_config_file)
QcConfigSignal(bigwig_cfg_df)
}
