% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcConfig.R
\name{QcConfigSignal}
\alias{QcConfigSignal}
\title{QcConfigSignal}
\usage{
QcConfigSignal(
  file_paths,
  groups = NULL,
  group_names = NULL,
  group_colors = NULL,
  view_size = 3000
)
}
\arguments{
\item{file_paths}{character paths to files}

\item{groups}{numeric vector of group assignments. 1 is first item in group_names, 2 is second, etc. Default is seq_along(file_path)}

\item{group_names}{vector of group names to assign from according to groups}

\item{group_colors}{vector of colors to use per group}

\item{view_size}{view size in bp to apply. Defaults to 3000.}
}
\value{
a QcConfigSignal object
}
\description{
QcConfigSignal
}
\examples{
QcConfigSignal(c("A", "B"))
}
