% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcConfig.R
\name{QcScaleFill}
\alias{QcScaleFill}
\alias{QcScaleFill,QcConfig-method}
\title{QcScaleFill}
\usage{
QcScaleFill(object)

\S4method{QcScaleFill}{QcConfig}(object)
}
\arguments{
\item{object}{a QcConfig object}
}
\value{
a ggplot2 scale_fill_manual
}
\description{
QcScaleFill

QcScaleFill for QcConfig
}
\examples{
df = data.frame(group = c("A", "B"))
df$x = 1:2
df$y = 1
library(ggplot2)
ggplot(df, aes(x = x, y = y, fill = group)) +
  geom_point(size = 20, pch = 22) +
  expand_limits(x = c(0, 3)) +
  QcScaleFill(QcConfig(df$group))
}
