% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tsne.R
\name{run_tsne}
\alias{run_tsne}
\title{Title}
\usage{
run_tsne(
  profile_dt,
  perplexity = 100,
  n_cores = getOption("mc.cores", 1),
  high_topright = TRUE,
  norm1 = TRUE,
  bfc = BiocFileCache::BiocFileCache(),
  rname = NULL,
  Y_init = NULL,
  verbose = TRUE,
  force_overwrite = FALSE,
  wide_var = "name",
  tall_var = "_none_"
)
}
\arguments{
\item{profile_dt}{}

\item{perplexity}{}

\item{n_cores}{}

\item{high_topright}{}

\item{norm1}{}

\item{bfc}{}

\item{rname}{}

\item{Y_init}{}

\item{verbose}{}

\item{force_overwrite}{}

\item{wide_var}{}

\item{tall_var}{}
}
\value{

}
\description{
Title
}
