% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssvQC.R
\name{ssvQC}
\alias{ssvQC}
\alias{ssvQC.runAll}
\alias{ssvQC.prepFragLens}
\alias{ssvQC.prepMappedReads}
\alias{ssvQC.prepCapValue}
\alias{ssvQC.plotMappedReads}
\alias{ssvQC.prepFRIP}
\alias{ssvQC.plotFRIP}
\alias{ssvQC.prepSCC}
\alias{ssvQC.plotSCC}
\alias{ssvQC.prepFetch}
\alias{ssvQC.referenceUsesSameScale}
\alias{ssvQC.prepSignal}
\alias{ssvQC.plotSignal}
\alias{ssvQC.prepFeatures}
\alias{ssvQC.plotFeatures}
\title{ssvQC}
\usage{
ssvQC(
  features_config = NULL,
  signal_config = NULL,
  out_dir = getwd(),
  bfc = NULL,
  matched_only = TRUE,
  ...
)

ssvQC.runAll(object)

ssvQC.prepFragLens(object, query, bfc)

ssvQC.prepMappedReads(object)

ssvQC.prepCapValue(object, query, bfc)

ssvQC.plotMappedReads(object)

ssvQC.prepFRIP(object)

ssvQC.plotFRIP(object)

ssvQC.prepSCC(object)

ssvQC.plotSCC(object)

ssvQC.prepFetch(object)

ssvQC.referenceUsesSameScale(object)

ssvQC.prepSignal(object)

ssvQC.plotSignal(object)

ssvQC.prepFeatures(object, bfc)

ssvQC.plotFeatures(object, force_euler)
}
\arguments{
\item{features_config}{Controls features configuration.  May be a:
QcConfigFeatures object, path to a file defining configuration via
QcConfigFeatures.parse, features files to define via
QcConfigFeatures.files, or a data.frame to pass to QcConfigFeatures.}

\item{signal_config}{Controls signal configuration.  May be a: QcConfigSignal
object, path to a file defining configuration via QcConfigSignal.parse,
features files to define via QcConfigSignal.files, or a data.frame to pass
to QcConfigSignal.}

\item{out_dir}{NYI}

\item{bfc}{BiocFileCache object to use for caching. If NULL, default
new_cache() will be used.}

\item{force_euler}{If TRUE forces Euler plots to be generated for a list of feature sets longer than 8.  Euler plots can take quite a long time to generate as more feature sets are generated.}
}
\value{
A ssvQC object.  Data needs to be loaded after via ssvQC.runAll or sub-methods ssvQC.plot*.
}
\description{
ssvQC
}
\examples{
options(mc.cores = 1)
set.seed(0)
features_config_file = system.file(package = "ssvQC", "extdata/ssvQC_peak_config.csv")
features_config = QcConfigFeatures.parse(features_config_file)

bam_config_file = system.file(package = "ssvQC", "extdata/ssvQC_bam_config.csv")
bam_config = QcConfigSignal.parse(bam_config_file)

bigwig_config_file = system.file(package = "ssvQC", "extdata/ssvQC_bigwig_config.csv")
bigwig_config = QcConfigSignal.parse(bigwig_config_file)

sqc.complete.file = ssvQC(features_config_file, bam_config_file)

sqc.complete = ssvQC(features_config, bam_config)
object = sqc.complete

sqc.complete.bw = ssvQC(features_config, bigwig_config_file)

sqc.signal = ssvQC(signal_config = bam_config)

sqc.feature = ssvQC(features_config = features_config)

sqc.signal = ssvQC.runAll(sqc.signal)
sqc.feature = ssvQC.runAll(sqc.feature)

sqc.complete = ssvQC.runAll(sqc.complete)

sqc.complete$plots$signal$heatmaps
sqc.complete$signal_config@plot_value = "RPM"
sqc.complete = ssvQC.plotSignal(sqc.complete)
sqc.complete$plots$signal$heatmaps

sqc.complete$signal_config@plot_value = "linearQuantile"
sqc.complete = ssvQC.plotSignal(sqc.complete)
sqc.complete$plots$signal$heatmaps

sqc.complete$signal_config@plot_value = SQC_SIGNAL_VALUES$RPM_linearQuantile
sqc.complete = ssvQC.plotSignal(sqc.complete)
sqc.complete$plots$signal$heatmaps
}
