% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssvQC.R
\name{ssvQC}
\alias{ssvQC}
\title{ssvQC}
\usage{
ssvQC(
  feature_config = NULL,
  signal_config = NULL,
  out_dir = getwd(),
  bfc = NULL
)
}
\arguments{
\item{feature_config}{}

\item{signal_config}{}

\item{out_dir}{}

\item{bfc}{}
}
\value{

}
\description{
ssvQC
}
\examples{
feature_config_file = system.file(package = "ssvQC", "extdata/ssvQC_peak_config.csv")
feature_config = QcConfigFeatures.parse(feature_config_file)

bam_config_file = system.file(package = "ssvQC", "extdata/ssvQC_bam_config.csv")
bam_config = QcConfigSignal.parse(bam_config_file)

bigwig_config_file = system.file(package = "ssvQC", "extdata/ssvQC_bigwig_config.csv")
bigwig_config = QcConfigSignal.parse(bigwig_config_file)

sqc.complete.file = ssvQC(feature_config_file, bam_config_file)

sqc.complete = ssvQC(feature_config, bam_config)

sqc.complete.bw = ssvQC(feature_config, bigwig_config_file)

sqc.signal = ssvQC(signal_config = bam_config)

sqc.feature = ssvQC(feature_config = feature_config)

ssvQC.runAll(sqc.complete)
ssvQC.runAll(sqc.signal)
ssvQC.runAll(sqc.feature)

sqc = sqc.complete
object = sqc

sqc = ssvQC.prepMappedReads(sqc)

p_reads = ssvQC.plotMappedReads(sqc)

sqc = ssvQC.prepSignal(sqc)

}
