% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssvQC.R
\name{ssvQC}
\alias{ssvQC}
\title{ssvQC}
\usage{
ssvQC(feature_config, signal_config, out_dir = getwd(), bfc = NULL)
}
\arguments{
\item{feature_config}{}

\item{signal_config}{}

\item{out_dir}{}

\item{bfc}{}
}
\value{

}
\description{
ssvQC
}
\examples{
feature_config_file = system.file(package = "ssvQC", "extdata/ssvQC_peak_config.csv")
feature_config = QcConfigFeatures.parse(feature_config_file)

bam_config_file = system.file(package = "ssvQC", "extdata/ssvQC_bam_config.csv")
bam_config = QcConfigSignal.parse(bam_config_file)

bigwig_config_file = system.file(package = "ssvQC", "extdata/ssvQC_bigwig_config.csv")
bigwig_config = QcConfigSignal.parse(bigwig_config_file)

ssvQC(feature_config_file, bam_config_file)

ssvQC(feature_config, bam_config)

ssvQC(feature_config, bigwig_config_file)

}
