% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssvQC_select.R
\name{ssvQC.removeFeatures}
\alias{ssvQC.removeFeatures}
\title{ssvQC.removeFeatures}
\usage{
ssvQC.removeFeatures(
  sqc,
  ids = NULL,
  grs = NULL,
  features_name = NULL,
  invert = FALSE
)
}
\arguments{
\item{sqc}{}

\item{ids}{}

\item{grs}{}

\item{invert}{}
}
\value{

}
\description{
ssvQC.removeFeatures
}
