% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssvQC.R
\name{ssvQC.save_config}
\alias{ssvQC.save_config}
\title{ssvQC.save_config}
\usage{
ssvQC.save_config(object, file)
}
\arguments{
\item{object}{A ssvQC object}

\item{file}{A .features.csv and .signal.csv file will be written with this prefix.}
}
\value{
invisibly returns config file paths.
}
\description{
ssvQC.save_config
}
\examples{
features_config_file = system.file(
  package = "ssvQC", 
  "extdata/ssvQC_peak_config.csv"
)
features_config = QcConfigFeatures.parse(features_config_file)

bam_config_file = system.file(
  package = "ssvQC", 
  "extdata/ssvQC_bam_config.csv"
)
bam_config = QcConfigSignal.parse(bam_config_file)

sqc = ssvQC(features_config, bam_config)

out_dir = tempdir()
ssvQC.save_config(sqc, file.path(out_dir, "test_save_config"))
}
