% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssvQC.R
\name{ssvQC.selectClusters}
\alias{ssvQC.selectClusters}
\title{Title}
\usage{
ssvQC.selectClusters(
  sqc,
  cluster_numbers,
  features_name = NULL,
  signals_name = NULL,
  invert = FALSE
)
}
\arguments{
\item{sqc}{}

\item{cluster_numbers}{}

\item{features_name}{}

\item{signals_name}{}

\item{invert}{}
}
\value{

}
\description{
Title
}
