% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssvQC.R
\name{ssvQC.selectFeatures}
\alias{ssvQC.selectFeatures}
\title{ssvQC.selectFeatures}
\usage{
ssvQC.selectFeatures(
  sqc,
  ids = NULL,
  grs = NULL,
  features_name = NULL,
  invert = FALSE
)
}
\arguments{
\item{sqc}{}

\item{ids}{}

\item{grs}{}

\item{features_name}{}

\item{invert}{}
}
\value{

}
\description{
ssvQC.selectFeatures
}
