% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssvTSNE.R
\name{ssvTSNE}
\alias{ssvTSNE}
\title{ssvTSNE}
\usage{
ssvTSNE(
  features_config = NULL,
  signal_config = NULL,
  out_dir = getwd(),
  bfc = NULL,
  matched_only = TRUE,
  n_glyphs_x = 10,
  n_glyphs_y = n_glyphs_x,
  n_heatmap_pixels_x = 25,
  n_heatmap_pixels_y = n_heatmap_pixels_x,
  ...
)
}
\arguments{
\item{features_config}{}

\item{signal_config}{}

\item{out_dir}{}

\item{bfc}{}

\item{matched_only}{}

\item{...}{}
}
\value{

}
\description{
ssvTSNE
}
\examples{
options(mc.cores = 1)
set.seed(0)
features_config_file = system.file(package = "ssvQC", "extdata/ssvQC_peak_config.csv")
features_config = QcConfigFeatures.parse(features_config_file)

bam_config_file = system.file(package = "ssvQC", "extdata/ssvQC_bam_config.csv")
bam_config = QcConfigSignal.parse(bam_config_file)
bam_config$view_size = 600

# TSNE plot paramters can be controlled at ssvTSNE object creation or set later
# if the y settings aren't specified they are the same as x
sts = ssvTSNE(
  features_config, 
  bam_config, 
  n_glyphs_x = 3, 
  n_heatmap_pixels_x = 5)
sts = ssvQC.plotFeatures(sts)
sts$perplexity = 10
sts = ssvQC.prepFetch(sts)
sts = ssvQC.referenceUsesSameScale(sts)
sts = ssvQC.prepSignal(sts)
sts = ssvQC.plotSignal(sts)
sts$plots$TSNE

sts$n_glyphs_x = 7 
sts$n_glyphs_y = 5

sts$n_heatmap_pixels_x = 30
sts$n_heatmap_pixels_y = 60
sts.replot = ssvQC.plotSignal(sts)

}
