% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssvTSNE.R
\name{ssvTSNE}
\alias{ssvTSNE}
\title{ssvTSNE}
\usage{
ssvTSNE(
  features_config = NULL,
  signal_config = NULL,
  out_dir = getwd(),
  bfc = NULL,
  matched_only = TRUE,
  ...
)
}
\arguments{
\item{features_config}{}

\item{signal_config}{}

\item{out_dir}{}

\item{bfc}{}

\item{matched_only}{}

\item{...}{}
}
\value{

}
\description{
ssvTSNE
}
\examples{
options(mc.cores = 1)
set.seed(0)
features_config_file = system.file(package = "ssvQC", "extdata/ssvQC_peak_config.csv")
features_config = QcConfigFeatures.parse(features_config_file)

bam_config_file = system.file(package = "ssvQC", "extdata/ssvQC_bam_config.csv")
bam_config = QcConfigSignal.parse(bam_config_file)
bam_config$view_size = 600

sts = ssvTSNE(features_config, bam_config)
sts = ssvQC.plotFeatures(sts)
sts = ssvQC.prepSignal(sts)
sts = ssvQC.plotSignal(sts)
sts$plots$signal
}
