% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_ssvQC.summary}
\alias{write_ssvQC.summary}
\title{write_ssvQC.summary}
\usage{
write_ssvQC.summary(sqc, out_dir = get_wd())
}
\arguments{
\item{sqc}{}

\item{out_dir}{}
}
\value{

}
\description{
write_ssvQC.summary
}
